Organize code, introduce eslint
