+++ examples/parlai_chat_task_demo/webapp/.eslintrc
@@ -1,0 +1,3 @@
+{
+  "extends": ["react-app"]
+}
