+++ packages/mephisto-task/package.json
@@ -18,6 +18,10 @@
     "@babel/preset-react": "^7.9.4",
     "babel-loader": "^8.1.0",
     "css-loader": "^3.5.1",
+    "eslint": "^7.0.0",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-react": "^7.20.0",
+    "eslint-plugin-react-hooks": "^4.0.2",
     "react": "16.13.1",
     "react-dom": "^16.13.1",
     "style-loader": "^1.1.3",
@@ -31,6 +35,7 @@
   },
   "dependencies": {
     "axios": "^0.19.2",
+    "babel-eslint": "^10.1.0",
     "bowser": "^2.9.0"
   }
 }
